+++ src/checks/lowercase_key.rs
@@ -43,39 +43,36 @@ mod tests {
 
     #[test]
     fn working_run() {
-        let mut checker = LowercaseKeyChecker::default();
-        let line = line_entry(1, 1, "FOO=BAR");
-        assert_eq!(None, checker.run(&line));
+        check_test(&mut LowercaseKeyChecker::default(), [("FOO=BAR", None)]);
     }
 
     #[test]
     fn working_with_export_run() {
-        let mut checker = LowercaseKeyChecker::default();
-        let line = line_entry(1, 1, "export FOO=BAR");
-        assert_eq!(None, checker.run(&line));
+        check_test(
+            &mut LowercaseKeyChecker::default(),
+            [("export FOO=BAR", None)],
+        );
     }
 
     #[test]
     fn failing_run_with_lowercase_key() {
-        let mut checker = LowercaseKeyChecker::default();
-        let line = line_entry(1, 1, "foo_bar=FOOBAR");
-        let expected = Some(Warning::new(
-            line.clone(),
-            "LowercaseKey",
-            "The foo_bar key should be in uppercase",
-        ));
-        assert_eq!(expected, checker.run(&line));
+        check_test(
+            &mut LowercaseKeyChecker::default(),
+            [(
+                "foo_bar=FOOBAR",
+                Some("The foo_bar key should be in uppercase"),
+            )],
+        );
     }
 
     #[test]
     fn failing_run_with_lowercase_letter() {
-        let mut checker = LowercaseKeyChecker::default();
-        let line = line_entry(1, 1, "FOo_BAR=FOOBAR");
-        let expected = Some(Warning::new(
-            line.clone(),
-            "LowercaseKey",
-            "The FOo_BAR key should be in uppercase",
-        ));
-        assert_eq!(expected, checker.run(&line));
+        check_test(
+            &mut LowercaseKeyChecker::default(),
+            [(
+                "FOo_BAR=FOOBAR",
+                Some("The FOo_BAR key should be in uppercase"),
+            )],
+        );
     }
 }